test: AuthControllerTest customGiven 적용
